The stocks are now ranked by low price over every day.
Fixed a bug where stocks without data for a day would make other stock go down in rank; missing days not use previous day data.

The ranks are now stored in a hashmap for faster retrieval.

The rank data is crudely graphed with matplotlib.

A date struct was created to share the date between structs.